첨부파일

+++ src/main/java/wam/app/w/asst/TAmAsstApiController.java
@@ -1,8 +1,10 @@
 package wam.app.w.asst;
 
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.slf4j.Logger;
@@ -12,18 +14,24 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
+import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
+import wam.app.util.CmFileUtils;
+import wam.app.w.file.TCmFile;
+import wam.app.w.file.TCmFileRepository;
 
 @Slf4j
 @Tag( name = "TAmAsstApiController", description = "자산")
@@ -37,6 +45,8 @@ public class TAmAsstApiController {
 	TAmAsstRepository asstRepository;
 	@Autowired
 	TAmAsstQuerydslRepository asstQuerydslRepository;
+	@Autowired
+	TCmFileRepository fileRepository;
 
 	@Operation(summary = "자산 목록 조회", description = "검색 값으로 페이징된 자산 목록 화면을 호출한다.")
 	@GetMapping("/assts")
@@ -55,22 +65,33 @@ public class TAmAsstApiController {
 	@GetMapping("/assts/{asstSn}")
 	public Optional<TAmAsst> get( @PathVariable Integer asstSn) throws Exception {
 
-		Optional<TAmAsst> asst = asstRepository.findById(asstSn);
+		Optional<TAmAsst> asstOpt = asstRepository.findById(asstSn);
 		
-		return asst;
+		return asstOpt;
 		
 	}
 	
 	@Operation(summary = "자산 저장", description = "자산 저장한다.")
-	@PutMapping("/assts")
-	public TAmAsst put(@RequestBody @Valid TAmAsst tWdTerm)  throws Exception  {
+	@RequestMapping(value="/assts", method=RequestMethod.PUT, consumes= {"multipart/form-data"})
+	public TAmAsst put(TAmAsst tAmAsst, HttpServletRequest request)  throws Exception  {
 		
-		log.debug("자산 저장 호출 : {}", tWdTerm);
+		log.debug("자산 저장 호출 : {}", tAmAsst);
 		TAmAsst asst;
-		asst = asstRepository.save(tWdTerm);
+		asst = asstRepository.save(tAmAsst);
 		if (asst == null) {
 			log.debug("저장시 오류발생");
+		} else {
+			// 파일저장 : vo로 입력받는 것이 아니라, request에서 값을 받아 저장
+			String tableNm = "T_AM_ASST";
+			String tableId = asst.getAsstSn().toString();
+			String path = tableNm;
+			log.debug("tableId : " + tableId);
+			List<TCmFile> fileList = CmFileUtils.saveMultiFile(tableNm,tableId,request,path); // 파일 저장
+			for(TCmFile vo :fileList) {
+				fileRepository.save(vo); // 파일정보 DB저장
+			}
 		}
+			
 		return asst;
 		
 	}
@@ -78,9 +99,12 @@ public class TAmAsstApiController {
 	@Operation(summary = "자산 삭제", description = "자산 삭제한다.")
 	@DeleteMapping("/assts/{asstSn}")
 	public String delete(@PathVariable Integer asstSn) throws Exception {
-		
 		log.debug("자산 삭제 호출 :"+  Integer.toString(asstSn));
+		if (asstSn == null || asstSn < 1) return "201"; // 자산 일련번호 오류
+
+		// 자산 삭제
 		asstRepository.deleteById(asstSn);
+		
 		return "200";
 		
 	}
